Moved handleRequestAsync to the superclass IntraHandlerRest. Now, each handler implements a function for each HTTP method.
